api_movies/requirements.txt
pydantic==2.12.0
pydantic-settings==2.11.0
fastapi[all]==0.111.0
uvicorn==0.38.0
sqlalchemy[asyncio]==2.0.44
asyncpg==0.30.0
PyJWT==2.10.1
aiohttp==3.12.15